requirements.txt
flask==2.3.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
Pillow==10.1.0
scikit-learn==1.3.2
numpy==1.24.4
opencv-python==4.8.1.78
gunicorn==21.2.0
python-multipart==0.0.6
colorthief==0.2.1
webcolors==1.13 